+++ app/models/webdav_step.rb
@@ -1,6 +1,14 @@
 # frozen_string_literal: true
 
 class WebdavStep < Step
+  attr_reader :run_modes
+
+  def initialize(name, run_modes, required_files = [])
+    super(name, required_files)
+
+    @run_modes = run_modes
+  end
+
   def builder_for(pipeline, params)
     PipelineSteps::Webdav::Builder.new(pipeline, name, params)
   end