+++ spec/factories/computations.rb
@@ -10,15 +10,16 @@ FactoryBot.define do
 
     factory :webdav_computation, class: 'WebdavComputation' do
       pipeline_step 'segmentation'
-      input_path { '/inputs' }
-      output_path { '/outputs' }
+      input_path '/inputs'
+      output_path '/outputs'
+      run_mode 'Workflow 3 (TEE Aortic Valve Segmentation)'
     end
 
     factory :rimrock_computation, class: 'RimrockComputation' do
       pipeline_step '0d_models'
       input_path nil
       output_path nil
-      script { 'SCRIPT' }
+      script 'SCRIPT'
       tag_or_branch 'master'
       revision '1234'
     end